Lens reader integration
To achieve Lens integration I have a separate build wrapper for lens: https://github.com/blahah/lens-starter
This includes a converter to deal with some funky aspects of EuPMC XML.

We run a content server which just static-serves files downloaded to the content store.
Lens is run inside a webview and calls out to the server to get content.

Currently the fulltext downloading only uses HTTP from EuPMC server.
There's a load of setup for using hyperdrive to get fulltexts that I will clean up and move to a new PR.